chore(table): refactor and clean up duplicate
